==> 01_2_type_cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns


#Open Files
file10 = "01_type_cleaning/model_freq_over_10.csv"
file200= "01_type_cleaning/model_freq_over_200.csv"
map = pd.read_csv(file200, delimiter = ";")
df = pd.read_csv(file10)

#Clean output Model and Manufcaturer
df['manufacturer'] = df['manufacturer'].str.lower()
df['model'] = df['model'].str.lower()
df.head(-5)

#Keep Only first 3 words of model
def keep_first_3_words(text):
    words = text.split()
    return ' '.join(words[:3]) if len(words) > 3 else text
#Keep Only first word of model
def keep_first_word(text):
    words = text.split()
    return ' '.join(words[:1]) if len(words) > 1 else text

#Add to Df
df['model_first3'] = df['model'].str.lower().apply(keep_first_3_words)
df['model_first'] = df['model'].str.lower().apply(keep_first_word)




#All Lowercase consistent mapping
map['manufacturer'] = map['manufacturer'].str.lower()
map['model_clean'] = map['model_clean'].str.lower()
map['manufacturer'] = map['manufacturer'].str.lower()

#Match model clean in model first 3
def match_model_clean(row, mapping):
    matches = mapping[mapping['manufacturer'] == row['manufacturer']]
    for _, m_row in matches.iterrows():
        if m_row['model_clean'] in row['model_first3']:
            return m_row['model_clean']
    return None

#Apply Matching
df['model_clean'] = df.apply(lambda row: match_model_clean(row, map), axis=1)


#If not first word in model Cleaned if NAN
df['model_clean'] = df['model_clean'].fillna(df['model_first'])

#Save Model Cleaned
df.to_csv("01_type_clean/Model Cleaned > 10.csv")

#One Last Check manual Check
