==> 02_encoding.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

file = "01_Used_Cars_Cleaned.csv"
df = pd.read_csv(file)

#Price - Target -> np.log1p
#price_segment - Target -> Ordinal Encoding

#Year - Drop
#Condition - Drop
#type - drop
#paint_color - drop
#type - drop
#paint_color - drop
#cylinders_clean - drop

#Manufacturer - frequency encoding -> np.log1p
#model - frequency encoding  -> np.log1p
#state - frequency encoding -> np.log1p

#odometer - np.log1p
#car_age - np.log1p
#miles_age - np.log1p

#model - One-Hot Encoding
#title_status - One-Hot Encoding
#transmission - One-Hot Encoding
#drive - One-Hot Encoding
# #grouped_color - One-Hot Encoding
#type_grouped - One-Hot Encoding
#Fuel - One-Hot Encoding

#condition_grouped - Ordinal Encoding
#cylinders_cat - Ordinal Encoding

#Delete Features
not_used = ['year', 'condition', 'type', 'paint_color', 'region']
df = df.drop(columns=not_used)

#Frequency Encoding
df['model_freq'] = df['model'].map(df['model'].value_counts())
df['state_freq'] = df['state'].map(df['state'].value_counts())
df['manufacturer_freq'] = df['manufacturer'].map(df['manufacturer'].value_counts())

df['log_model_freq'] = np.log1p(df['model_freq'])
df['log_state_freq'] = np.log1p(df['state_freq'])

dropped = ['model', 'state','model_freq','state_freq']
df = df.drop(columns=dropped)

#Log Transform
df['log_odometer'] = np.log1p(df['odometer'])
df['log_car_age'] = np.log1p(df['car_age'])
df['log_miles_age'] = np.log1p(df['miles_age'])

drops = ['odometer', 'car_age', 'miles_age']
df = df.drop(columns=drops)

#Ordinal Encoding
condition_map = {'salvage': 1, 'fair': 2, 'good': 3, 'excellent': 4}
df['condition_grouped'] = df['condition_grouped'].map(condition_map)

cylinders_map = {'low': 1, 'mid': 2, 'high': 3}
df['cylinders_cat'] = df['cylinders_cat'].map(cylinders_map)

#Manufacturer Ordinal Encoding based on Market Capitalization
manufacturer_marketcap_map = {
    "toyota": 1,
    "mercedes-benz": 2,
    "volkswagen": 3,
    "bmw": 4,
    "porsche": 5,
    "tesla": 6,
    "honda": 7,
    "ford": 8,
    "volvo": 9,
    "audi": 10,
    "datsun": 11,
    "mitsubishi": 11,
    "nissan": 11,
    "chevrolet": 12,
    "hyundai": 13,
    "lexus": 14,
    "subaru": 15,
    "land rover": 16,
    "jeep": 17,
    "kia": 18,
    "buick": 19,
    "cadillac": 20,
    "mazda": 21,
    "gmc": 22,
    "ram": 23,
    "mini": 24,
    "jaguar": 25,
    "fiat": 26,
    "lincoln": 27,
    "acura": 28,
    "harley-davidson": 29,
    "dodge": 30,
    "infiniti": 31,
    "chrysler": 32,
    "alfa-romeo": 33,
    "mercury": 99,
    "pontiac": 99,
    "saturn": 99
}

df['manufacturer_marketcap'] = df['manufacturer'].map(manufacturer_marketcap_map)

#Encoding Target Variables
price_segment_map= {'low': 1, 'mid': 2, 'high': 3}
df['price_segment'] = df['price_segment'].map(price_segment_map)

df['log_price'] = np.log1p(df['price'])
df = df.drop(columns='price')

#Save Document
df.to_csv("02_Used_Cars_Pipeline.csv", index=False)
